Set defaults for trivial values.
This will set default values for the variables edit_map_id and scroll_x
and scroll_y for maps. While the default edit_map_id is set in
config.yaml (and can be overwritten with -1) the scroll settings for
MapInfos will always be set to 0 for both scroll_x and scroll_y.

The reason this change is done, is to avoid issues with merge conflicts,
just because a user had one map scrolled to one location, whilest
another had the scrollbar 1 px further to the right.

The issue with order still remains for maps, which probably can't be
solved without doing some modifications to the data structure.